# docs/openapi.yaml
openapi: 3.0.0
info:
  title: IntegraHub API
  description: Enterprise Integration Platform API
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Gateway URL

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

security:
  - bearerAuth: []

paths:
  /auth/login:
    post:
      summary: Authenticate and get JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  example: password
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
  
  /orders:
    post:
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                customer_id:
                  type: string
                  example: "cust-001"
                total_amount:
                  type: number
                  example: 150.50
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      name: 
                        type: string
                      quantity:
                        type: integer
                      price:
                        type: number
      responses:
        '201':
          description: Order Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                   type: string
                  order_id:
                   type: integer
                  correlation_id:
                   type: string
        '401':
          description: Unauthorized
  
  /orders/{id}:
    get:
      summary: Get order details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Order details
        '404':
          description: Order not found

  /analytics:
    get:
      summary: Get system analytics
      responses:
        '200':
          description: Metrics
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_events:
                    type: integer
                  events_by_type:
                    type: object